Small changes, rename configgg to DatasourceConfig
