+++ build.gradle
@@ -44,7 +44,6 @@ tasks.withType(com.github.spotbugs.SpotBugsTask) {
 
 
 dependencies {
-    compile group: 'org.postgresql', name: 'postgresql', version: '42.2.10'
     compileOnly 'com.github.spotbugs:spotbugs-annotations:4.0.0-RC1'
     implementation 'mops:styleguide:2.1.0'
     testCompile("org.assertj:assertj-core:3.15.0")
@@ -65,6 +64,7 @@ dependencies {
     testCompile("org.assertj:assertj-core:3.11.1")
     compile 'org.thymeleaf.extras:thymeleaf-extras-springsecurity5:3.0.4.RELEASE'
     implementation 'junit:junit:4.12'
+    compile group: 'org.postgresql', name: 'postgresql', version: '42.2.10'
 }
 
 test {
