--- src/main/java/mops/config/configgg.java
@@ -1,57 +1,0 @@
-package mops.config;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.env.Environment;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-
-import javax.sql.DataSource;
-import java.sql.Connection;
-import java.sql.DatabaseMetaData;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-import java.util.Properties;
-
-@Configuration
-public class configgg {
-    @Autowired
-    private Environment env;
-
-    @Bean
-    public void connect() {
-        try {
-            Class.forName("org.postgresql.Driver");
-            String dbURL = "jdbc:postgresql://db:5432/moepsinnen";
-            Properties props = new Properties();
-            props.setProperty("user", "mops");
-            props.setProperty("password", "mops");
-            props.setProperty("ssl", "false");
-            Connection conn = DriverManager.getConnection(dbURL, props);
-            if (conn != null) {
-                System.out.println("Connected to the database");
-                DatabaseMetaData dm = conn.getMetaData();
-                System.out.println("Driver name: " + dm.getDriverName());
-                System.out.println("Driver version: " + dm.getDriverVersion());
-                System.out.println("Product name: " + dm.getDatabaseProductName());
-                System.out.println("Product version: " + dm.getDatabaseProductVersion());
-                conn.close();
-            }
-        } catch (ClassNotFoundException ex) {
-            ex.printStackTrace();
-        } catch (SQLException ex) {
-            ex.printStackTrace();
-        }
-    }
-
-    @Bean
-    public DataSource dataSource() {
-        final DriverManagerDataSource dataSource = new DriverManagerDataSource();
-        dataSource.setDriverClassName("org.postgresql.Driver");
-        dataSource.setUrl(env.getProperty("jdbc.url"));
-        dataSource.setUsername(env.getProperty("jdbc.user"));
-        dataSource.setPassword(env.getProperty("jdbc.password"));
-        return dataSource;
-    }
-
-}
