+++ src/main/resources/application.properties
@@ -3,7 +3,6 @@ spring.jpa.show-sql=false
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.profiles.active=web
 spring.datasource.initialization-mode=always
-spring.datasource.driver-class-name=org.postgresql.Driver
 spring.datasource.password=mops
 spring.datasource.username=mops
 spring.datasource.url=jdbc:postgresql://db:5432/moepsinnen
